// frontend/src/api/models/createAiGameDtoSide.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Warcabownik
 * Checkers model with database and API for managing games
 * OpenAPI spec version: 1.0
 */

/**
 * Side the human player wants to play as. Defaults to WHITE when omitted.
 */
export type CreateAiGameDtoSide = typeof CreateAiGameDtoSide[keyof typeof CreateAiGameDtoSide];


export const CreateAiGameDtoSide = {
  WHITE: 'WHITE',
  BLACK: 'BLACK',
} as const;